Fix handling of zero-length label written last in the lexicon
There is no restriction in place disallowing zero-length labels, and in some use cases,
it makes sense to have them. However, when the lexicon file happens to include a
zero-length label as the last label in the list, split(" ") swallows it. This can later lead
to ArrayIndexOutOfBoundExceptions when something gets classified with the
zero-length label.